+++ common/scripts/definitions/objects/buildings_base.ts
@@ -680,8 +680,8 @@ export function Buildings_Default_Init(buildings:Definitions<BuildingDef,{}>){
             content:[
                 //{def:"md_crate",position:v2.zero},
                 {def:"airdrop_locked",position:v2.zero},
-                {def:"barrel",position:v2(1.7,1.7)},
 
+                //{def:"barrel",position:v2(1.7,1.7)},
                 //{def:"metal_door",position:v2(-2.5,-0.7),rotation:1,variation:7}
             ]
         }),