+++ server/src/game_server/objects/parachute.ts
@@ -57,7 +57,7 @@ export class Parachute extends ServerGameObject{
             }
 
             const def=this.game.definitions.synced_particle.getFromString("airdrop_smoke")
-            for(let i=0;i<8;i++){
+            for(let i=0;i<6;i++){
                 this.game.add_synced_particle(this.position,def,undefined,this.layer)
             }
         }
@@ -66,7 +66,7 @@ export class Parachute extends ServerGameObject{
     }
     create(args: {position:Vec2,obstacle:ObstacleDef}): void {
         this.parachute_data={
-            lifetime:15,
+            lifetime:10,
             spawn_obstacle:args.obstacle
         }
         this.base_hitbox=new CircleHitbox2D(v2.zero(),3)